\name{voileta-package}
\alias{voileta-package}
\alias{voileta}
\docType{package}
\title{
\packageTitle{voileta}
}
\description{
\packageDescription{voileta}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{voileta}
\packageIndices{voileta}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{voileta}

Maintainer: \packageMaintainer{voileta}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
